==> code/Notebook 4 - Statistical Modelling.ipynb <==
# Importing required libraries and packages
import pandas as pd
import numpy as np
import geopandas as gpd
import warnings
import matplotlib.pyplot as plt

from collections import defaultdict
from sklearn.neural_network import MLPClassifier
from statsmodels.formula.api import *

warnings.filterwarnings("ignore")

# Loads in the preprocessed data
# TLC
yellow_2019 = pd.read_csv("../preprocessed_data/yellow_2019.csv")
yellow_2020 = pd.read_csv("../preprocessed_data/yellow_2020.csv")
green_2019 = pd.read_csv("../preprocessed_data/green_2019.csv")
green_2020 = pd.read_csv("../preprocessed_data/green_2020.csv")
zone = pd.read_csv("../raw_data/taxi+_zone_lookup.csv")

# Unemployment
unemployment = pd.read_csv('../preprocessed_data/unemployment.csv')

conditions = [
    (yellow_2019['tip_amount'] <= 2),
    (yellow_2019['tip_amount'] > 2) & (yellow_2019['tip_amount'] <= 5),
    (yellow_2019['tip_amount'] > 5) & (yellow_2019['tip_amount'] <= 10),
    (yellow_2019['tip_amount'] > 10) & (yellow_2019['tip_amount'] <= 100),
    (yellow_2019['tip_amount'] > 100)]

choices = [2,5,10,100,200]

yellow_2019['bin'] = np.select(conditions, choices)

conditions = [
    (yellow_2020['tip_amount'] <= 2),
    (yellow_2020['tip_amount'] > 2) & (yellow_2020['tip_amount'] <= 5),
    (yellow_2020['tip_amount'] > 5) & (yellow_2020['tip_amount'] <= 10),
    (yellow_2020['tip_amount'] > 10) & (yellow_2020['tip_amount'] <= 100),
    (yellow_2020['tip_amount'] > 100)]

choices = [2,5,10,100,200]

yellow_2020['bin'] = np.select(conditions, choices)

conditions = [
    (green_2019['tip_amount'] <= 2),
    (green_2019['tip_amount'] > 2) & (green_2019['tip_amount'] <= 5),
    (green_2019['tip_amount'] > 5) & (green_2019['tip_amount'] <= 10),
    (green_2019['tip_amount'] > 10) & (green_2019['tip_amount'] <= 100),
    (green_2019['tip_amount'] > 100)]

choices = [2,5,10,100,200]

green_2019['bin'] = np.select(conditions, choices)

conditions = [
    (green_2020['tip_amount'] <= 2),
    (green_2020['tip_amount'] > 2) & (green_2020['tip_amount'] <= 5),
    (green_2020['tip_amount'] > 5) & (green_2020['tip_amount'] <= 10),
    (green_2020['tip_amount'] > 10) & (green_2020['tip_amount'] <= 100),
    (green_2020['tip_amount'] > 100)]

choices = [2,5,10,100,200]

green_2020['bin'] = np.select(conditions, choices)

fit = ols(formula="tip_amount ~ fare_amount + borough + trip_distance + unemployment + total_amount",
         data=yellow_2019).fit()
print(fit.summary())

fit = ols(formula="tip_amount ~ fare_amount + borough + trip_distance + unemployment + total_amount",
         data=yellow_2020).fit()
print(fit.summary())

fit = ols(formula="tip_amount ~ fare_amount + borough + trip_distance + unemployment + total_amount",
         data=green_2019).fit()
print(fit.summary())

fit = ols(formula="tip_amount ~ fare_amount + borough + trip_distance + unemployment + total_amount",
         data=green_2020).fit()
print(fit.summary())

X_train = green_2019.loc[:, green_2019.columns != 'bin']
X_train = X_train.loc[:, X_train.columns != 'tip_amount'].to_numpy()
X_test = green_2020.loc[:, green_2020.columns != 'bin']
X_test = X_test.loc[:, X_test.columns != 'tip_amount'].to_numpy()
y_train = green_2019.loc['bin'].to_numpy()
y_test = green_2020.loc['bin'].to_numpy()
clf = MLPClassifier().fit(X_train, y_train)
clf.score(X_test, y_test)

X_train = yellow_2019.loc[:, yellow_2019.columns != 'bin']
X_train = X_train.loc[:, X_train.columns != 'tip_amount'].to_numpy()
X_test = yellow_2020.loc[:, yellow_2020.columns != 'bin']
X_test = X_test.loc[:, X_test.columns != 'tip_amount'].to_numpy()
y_train = yellow_2019.loc['bin'].to_numpy()
y_test = yellow_2020.loc['bin'].to_numpy()
clf = MLPClassifier().fit(X_train, y_train)
clf.score(X_test, y_test)
